+++ src/components/sidebar/Sidebar.jsx
@@ -19,13 +19,9 @@ import {
   SiYoutube,
 } from 'react-icons/si'
 import { useDispatch, useSelector } from 'react-redux'
-import {
-  Link as ReactRouterLink,
-  useLocation,
-  useNavigate,
-} from 'react-router-dom'
+import { Link as ReactRouterLink, useNavigate } from 'react-router-dom'
 
-import { AccountsAPI, LobbiesAPI } from '@api'
+import { AccountsAPI } from '@api'
 import logoFull from '@assets/images/logo.svg'
 import logoSymbol from '@assets/images/logo_symbol_white.svg'
 import {
@@ -49,8 +45,7 @@ import {
   UserIcon,
 } from '@components'
 import { StorageService } from '@services'
-import { addToast, toggleFriendList } from '@slices/AppSlice'
-import { updateLobby } from '@slices/LobbySlice'
+import { toggleFriendList } from '@slices/AppSlice'
 import { updateUser } from '@slices/UserSlice'
 
 import style from './Sidebar.module.css'
@@ -67,14 +62,12 @@ export default function Sidebar({ collapsed = true, collapsable = false }) {
 
   const dispatch = useDispatch()
   const navigate = useNavigate()
-  const location = useLocation()
 
   const [isCollapsed, setIsCollapsed] = useState(collapsable && collapsed)
   const [openSupport, setOpenSupport] = useState(false)
   const [notificationsOpen, setNotificationsOpen] = useState(false)
   const [friendListOpen, setFriendListOpen] = useState(false)
   const [unreadNotificationsCount, setUnreadNotificationsCount] = useState(0)
-  const [isFetching, setIsFetching] = useState(false)
 
   const handleOpenModalSupport = () => {
     setOpenSupport(true)
@@ -135,14 +128,10 @@ export default function Sidebar({ collapsed = true, collapsable = false }) {
             variant="queue"
           >
             <Text w={10}>
-              {isFetching ? (
-                '...'
-              ) : (
-                <Timer
-                  initialTime={lobby.queue_time === 0 ? 1 : lobby.queue_time}
-                  stop={preMatch}
-                />
-              )}
+              <Timer
+                initialTime={lobby.queue_time === 0 ? 1 : lobby.queue_time}
+                stop={preMatch}
+              />
             </Text>
           </Button>
         )}
@@ -230,27 +219,6 @@ export default function Sidebar({ collapsed = true, collapsable = false }) {
       setUnreadNotificationsCount(notificationsNotRead)
     }
   }, [notifications])
-
-  useEffect(() => {
-    const fetch = async () => {
-      setIsFetching(true)
-      const userToken = StorageService.get('token')
-      const response = await LobbiesAPI.detail(userToken, user.lobby_id)
-      if (response.errorMsg) {
-        dispatch(
-          addToast({
-            content: response.errorMsg,
-            variant: 'error',
-          })
-        )
-        return
-      } else dispatch(updateLobby(response))
-      setIsFetching(false)
-    }
-
-    fetch()
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [user, location.pathname])
 
   return (
     <>