+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Adicionado `useEffect` no `App` para setar valor inicial no slice de `lobby` [#316](https://github.com/3C-gg/reload-frontend/issues/316).
 - Adicionado variante `pin` para o `Button`.
 - Adicionado case de `lobbies/update` no `WebSocket Service`.
 - Adicionado useEffect na `SideBar` para atualizar o slice de `lobby` toda vez que mudar página.
@@ -32,6 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Alterado prop `otherPlayers` no componente `LobbyLineup` para iniciar com `[]`.
 - Alterado estilo do `IconButton` e do `FormHelperText` da página de `Verify` para melhorar o visual [#310](https://github.com/3C-gg/reload-frontend/issues/310).
 - Alterado view `LobbyView` para adicionar ação no botão de `jogar` [#302](https://github.com/3C-gg/reload-frontend/issues/302).
 - Alterado `SideBar` para usar o `lobby` do próprio slice de `lobby`.
@@ -53,6 +55,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+- Removido reducer `initLobby` de `LobbySlice` pois é redundante com o reducer `updateLobby`.
+- Removido `useEffect` que faz fetch de lobby `detail` no componente `SideBar`.
+- Removido `useEffect` que faz fetch de lobby `detail` na view `LobbyView`.
+- Removido view que não está sendo usada `BaseView` [#311](https://github.com/3C-gg/reload-frontend/issues/311). 
 - Removido antiga view de `Lobby`.
 - Removido componentes que não estão sendo utilizados `SteamSignin`, `LobbyHeader`, `FriendListUser`, 
 `InviteIstItem`, `UserCard`, `UserCardMini`, `UserStatus` [#253](https://github.com/3C-gg/reload-frontend/issues/253).